Use a slab cache for fdtable struct
We sometimes see slowdowns when running the threaded context switch test
from will-it-scale on a large ppc64le box. This was narrowed down to
excessive time in __fget_files, specifically the load of fdt->max_fds.

We allocate the fdtable struct directly from kmalloc, so it could share
the cacheline with any other small kmalloc. While the fdtable struct
is mostly read, the other allocation may not be.

Switching it to a slab cache fixes the issue.

Signed-off-by: Anton Blanchard <anton@ozlabs.org>