microwatt: Update liteeth dts for 16 tx/ 16 rx
Experiment with a large tx and rx ring. This needs a modification to
liteeth:

https://github.com/enjoy-digital/liteeth/pull/71

Then in microwatt add the following to the arty.yml or nexys-video.yml:

ntxslots:   16
nrxslots:   16

Then regenerate:

./generate.sh

Unfortunately the CSR layout changed in upstream liteeth so this patch
includes some shuffling around of the register layout.